test(🧪): run the SkMatrix scale spec against native too
Move it to renderer/__tests__/e2e and go through surface.eval, so the
JSI bindings this fixes are actually exercised. As a plain unit test it
only ever ran against CanvasKit, which was already correct.
